+++ src/main/java/com/megazone/ERPSystem_phase2_Backend/production/repository/basic_data/Workcenter/WorkcenterRepository.java
@@ -1,4 +1,4 @@
-package com.megazone.ERPSystem_phase2_Backend.production.repository.basic_information.Workcenter;
+package com.megazone.ERPSystem_phase2_Backend.production.repository.basic_data.Workcenter;
 
 import com.megazone.ERPSystem_phase2_Backend.production.model.basic_data.Workcenter;
 import org.springframework.data.jpa.repository.JpaRepository;
